+++ src/cli/run.rs
@@ -1,7 +1,4 @@
-use std::{
-    collections::{HashMap, HashSet},
-    path::PathBuf,
-};
+use std::{collections::HashMap, path::PathBuf};
 
 use chrono::NaiveDate;
 use clap::Parser;
@@ -147,23 +144,19 @@ fn issues_to_create_for_name<'a>(
 ) -> impl Iterator<Item = (u32, CreateIssuePayload)> + 'a {
     let current_issue = issue.most_recent_issue(date);
 
-    let already_issued = ledger
-        .get(name)
-        .map(|entries| entries.keys().copied().collect())
-        .unwrap_or_default();
+    ledger
+        .missing_issues(name, current_issue)
+        .into_iter()
+        .map(|issue_number| {
+            let payload = CreateIssuePayload {
+                project_path: issue.project.clone(),
+                description: Some(render(issue.template())),
+                due: Some(issue.due_date(issue_number)),
+                title: name.to_string(),
+            };
 
-    let to_issue = missing_issue_numbers(&already_issued, current_issue);
-
-    to_issue.into_iter().map(|issue_number| {
-        let payload = CreateIssuePayload {
-            project_path: issue.project.clone(),
-            description: Some(render(issue.template())),
-            due: Some(issue.due_date(issue_number)),
-            title: name.to_string(),
-        };
-
-        (issue_number, payload)
-    })
+            (issue_number, payload)
+        })
 }
 
 fn render(template: &str) -> String {
@@ -172,31 +165,4 @@ fn render(template: &str) -> String {
     templates
         .render(template, &tera::Context::default())
         .unwrap()
-}
-
-/// Find all missing numbers up to and including the current issue that have not already been issued.
-///
-/// 'already issued' is an iterator over values already issued, in ascending order.
-fn missing_issue_numbers(already_issued: &HashSet<u32>, current_issue: Option<u32>) -> Vec<u32> {
-    current_issue.map_or_else(Vec::default, |current_issue| {
-        (0..=current_issue)
-            .filter(|&n| !already_issued.contains(&n))
-            .collect()
-    })
-}
-
-#[cfg(test)]
-mod tests {
-
-    #[test]
-    fn missing_issue_numbers() {
-        let current_issue = Some(5);
-        let already_issued = [0, 1, 3, 5];
-        let expected = vec![2, 4];
-
-        let actual =
-            super::missing_issue_numbers(&already_issued.into_iter().collect(), current_issue);
-
-        assert_eq!(expected, actual);
-    }
 }
